``` ------ The CudaRandom pseudorandom seed is 0 Starting timing for: iteration 0 Winrate is 254 and 746 out of 1000. The time was 00:00:04.9877969 for: iteration 0 Starting timing for: iteration 1 Winrate is 444 and 556 out of 1000. The time was 00:00:07.1227760 for: iteration 1 Starting timing for: iteration 2 Winrate is 573 and 427 out of 1000. The time was 00:00:07.5869046 for: iteration 2 Starting timing for: iteration 3 Winrate is 643 and 357 out of 1000. The time was 00:00:07.5557583 for: iteration 3 Starting timing for: iteration 4 Winrate is 655 and 345 out of 1000. The time was 00:00:07.2145239 for: iteration 4 Starting timing for: iteration 5 Winrate is 661 and 339 out of 1000. The time was 00:00:07.3232094 for: iteration 5 Starting timing for: iteration 6 Winrate is 701 and 299 out of 1000. The time was 00:00:07.8085287 for: iteration 6 Starting timing for: iteration 7 Winrate is 704 and 296 out of 1000. The time was 00:00:07.3496127 for: iteration 7 Starting timing for: iteration 8 Winrate is 700 and 300 out of 1000. The time was 00:00:07.5648947 for: iteration 8 Starting timing for: iteration 9 Winrate is 700 and 300 out of 1000. The time was 00:00:07.7975557 for: iteration 9 Starting timing for: iteration 10 Winrate is 717 and 283 out of 1000. The time was 00:00:07.9319314 for: iteration 10 Starting timing for: iteration 11 Winrate is 747 and 253 out of 1000. The time was 00:00:07.7464394 for: iteration 11 Starting timing for: iteration 12 Winrate is 711 and 289 out of 1000. The time was 00:00:07.6660745 for: iteration 12 Starting timing for: iteration 13 Winrate is 715 and 285 out of 1000. The time was 00:00:07.7142631 for: iteration 13 Starting timing for: iteration 14 Winrate is 729 and 271 out of 1000. The time was 00:00:07.5218511 for: iteration 14 ```
This is actually pretty good.

```
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 275 and 725 out of 1000.
The time was 00:00:07.1075244 for: iteration 0
Starting timing for: iteration 1
Winrate is 523 and 477 out of 1000.
The time was 00:00:10.0481105 for: iteration 1
Starting timing for: iteration 2
Winrate is 622 and 378 out of 1000.
The time was 00:00:10.9522746 for: iteration 2
Starting timing for: iteration 3
Winrate is 670 and 330 out of 1000.
The time was 00:00:10.8196539 for: iteration 3
Starting timing for: iteration 4
Winrate is 681 and 319 out of 1000.
The time was 00:00:10.6150230 for: iteration 4
Starting timing for: iteration 5
Winrate is 704 and 296 out of 1000.
The time was 00:00:11.1955030 for: iteration 5
Starting timing for: iteration 6
Winrate is 720 and 280 out of 1000.
The time was 00:00:11.1668213 for: iteration 6
Starting timing for: iteration 7
Winrate is 718 and 282 out of 1000.
The time was 00:00:10.2784193 for: iteration 7
Starting timing for: iteration 8
Winrate is 719 and 281 out of 1000.
The time was 00:00:10.0544325 for: iteration 8
Starting timing for: iteration 9
Winrate is 690 and 310 out of 1000.
The time was 00:00:11.4485074 for: iteration 9
Starting timing for: iteration 10
Winrate is 726 and 274 out of 1000.
The time was 00:00:10.9228213 for: iteration 10
Starting timing for: iteration 11
Winrate is 732 and 268 out of 1000.
The time was 00:00:11.0001069 for: iteration 11
Starting timing for: iteration 12
Winrate is 687 and 313 out of 1000.
The time was 00:00:10.1090456 for: iteration 12
Starting timing for: iteration 13
Winrate is 703 and 297 out of 1000.
The time was 00:00:10.9773238 for: iteration 13
Starting timing for: iteration 14
Winrate is 719 and 281 out of 1000.
The time was 00:00:10.5610379 for: iteration 14
```

It could be a bit better, but compared to the results I was getting previously, this is in a league of its own.

The changed that I did compared to last time are:
1) Gather the gradients across timesteps rather than doing online updates.
2) Added a bias term.

Apart from that there is no real difference. Let me lower the learning rate a tad again and I'll see whether I get better asymptotic performance.