It is only halfway, but let me see.
...It is generally less stable than MC, but I get the feeling that unlike before, now it actually wants to work.

I could try running with a lower learning rate, but I can tell that it's biased regardless from that single run. It is not going to do as well as MC. With these results I do not have a reason to ever not use MC. And I do not think that TD is the answer to credit assignment over very long distances anyway.

Yeah, I am done with this. Good work me.