It is only halfway, but let me see.
...It is generally less stable than MC, but I get the feeling that unlike before, now it actually wants to work.

I could try running with a lower learning rate, but I can tell that it's biased regardless from that single run. It is not going to do as well as MC. With these results I do not have a reason to ever not use MC. And I do not think that TD is the answer to credit assignment over very long distances anyway.

Yeah, I am done with this. Good work me.

2:50pm. I've reached a milestone, so I think I will take a break here. My pace was just too breakneck over the past week.

Let me once more gather my thoughts. What I have I attained over the past two weeks.

1) I completely and thoroughly testing out PRONG a RL context.
2) I created na AC player.
3) I tested it with both MC and TD returns.
4) I discovered that TD works like crap and MC is great.
5) There is absolutely no reason to use a raw PG player over a AC with a MC critic.

I did an amazing amount of programming and testing in general. It is really a pity that TD did not work.

I expected it might not, but that cannot be helped. I guess a supervised component is necessary in the loop somewhere. And if I had to do it in a supervised fashion I'd rather just use MC rather than a target network in order to stabilize TD.

Though TD is popular in RL circles, it is not a real solution to doing credit assignment over long time horizons and it is true that at least in the tabular case it would take exponential time to do the bias correction. I won't be losing anything by not using it.

I am satisfied by the degree of effort I put into making this work. Now I can be sure where things stand.

One thing I will take away from the Zap paper is the `** 0.85` heuristic for updating the covariance and the steady state matrix. I've had good luck in applying that to the learning rate for the critic as well. It seems it is good when the critic updates faster than the main net as can be expected.

Right now as I write this I am playing with AC agent. It does seem that deeper architectures require a lower learning rate.

For the 2 layer one, the `2 ** -11.25` seems to be in the ideal range. This is in contrast to when I only trained the top layer which did well with `2 ** -9`. Think about learning rates in exponentials is another new habit I acquired. It is a lot better to describe them in powers of two rather than spell out the decimals like `0.001` and such.

Also it seems simpler architectures work better. The two layer net is worse than shallower ones. I expect this to change when I move to no limit holdem and self play. At that point the game will be complex enough that the extra capacity will be able to be taken advantage of.