As per plan, I think for the rest of the day I will just gather my thoughts like I did in the last commit.
The next goal would definitely be RNNs. When I tried them months ago with raw PG, it worked like crap, so I am going to have to take a look at whether the combination of PRONG and the MC critic is enough to get it into the stable range.

I might do well to revisit the recurrent K-FAC paper again and try and figure out what it is doing.

I might also bring back the LN+Relu activation, but it is not a priority.

Honestly, I've benefited quite a bit from simplyifying PRONG to the level where it is usable. I think that studying the recurrent K-FAC and higher order methods might make a worthwhile investment of my time.

So it goes roughly like this.

It does not feel good to me to just assume that K-FAC made for feedforward nets that I have now will be enough to do it.

To be 100% sure, I should master the recurrent variant, and pick up that thread on the tridiagonal one.

This means more time spent on studying. Yeah, yeah I know. But this is something I believe in and is of real importance in RL.

Nobody is really using this stuff apart from the OpenAI guys working on ACKTR. In the latest Dota 2 bout, the LSTM they are using is just plain PPO. I am guessing another group from within OpenAI is doing that.

I think that in the worst case, the recurrent K-FAC should be enough to get RNNs fully functional in the RL domain.

At that point I will have advanced in my education and will have another neglected asset in my toolbox.

That will put me in a prime position to tackle Rudder. After I get that I will be absolute.

The 'make an agent' phase will be done with. At that point the goal will change to 'scale them' and 'make use of them in the real world' after that.

It would be good if I could get through both RNN and Rudder in the next two months. I know that this will take me a while as usual.

Since it is 4pm and I definitely do not feel like programming again today, what I will do is pick up the recurrent K-FAC paper and do some studying.

As an aside, both the recurrent K-FAC and the tridiagonal update use Bayesian reasoning to get at the updates. At this point, rather than ask Reddit, I am thinking that I will try asking Martens directly if I really need to.