Ah, I do not get it. Well, I get that the K-FAC with the independence assumption is just standard K-FAC + division of the gradient by the number of timesteps. The other two approximation are way over my head. I basically cannot follow the math for those.
I have no idea whether I should give the Zap updates for RNNS a try, but I guess I could try. The worst that could happen is that I waste a day doing it.

Comparing the independent K-FAC to the other ones does not show it to be that far behind. I think I'll survive somehow regardless.