I am up. Despite going to bed a little later than usual, I really slept well tonight. For the past week I've been incredibly obsessed with finishing the AC agent, so now that is done it seems I am getting some relief from the demon driving me.
Now, if I started at the current I could bring in the recurrent MI layer in an hour or two. It is no biggie.

What is a biggie is the math in the K-FAC papers which I still do not understand. I've been thinking about the indenpendent K-FAC for recurrent nets and have concluded that that way I am doing it now is just fine. I do not understand the way why the normalization constant is done the way it is in the paper, but it is really not that important. Ultimately it should boil down to the way the covariance matrices are updated. I've gone from the full NG update and concluded that the way I am doing it now is fine.

For now, how about I read the original K-FAC paper again and then I guess I'll study Gaussian processes or something a bit more. I need to go back and read what /u/bbsome wrote to me.