"9:50am. 10/21. The stuff on this page could be my reason for trying out the Zap updates. It seems that rather than using h which is the identity, which I've been using up to now, I could potentially replace it with `(I - u' u^-1)`. That would give me the Zap update.
10:15am. It is an interesting, but convoluted paper as all NG papers are. I skipped the stuff on pruning. But basically I know now that by using the Zap update, I am simply assuming a different distance metric. It is definitely worth trying out. Something simple like Zap will definitely be a boon over the standard update for RNNs.

It does not seem like much, but I did gain quite a bit from reading these papers. Until I read the RNN K-FAC paper I actually was not aware of the independent update. I read the paper before, but I did not internalize anything of it until now.

28/1098. This is Kevin Murphy's ML book that I am going through now. I guess I'll dedicate today and maybe a few days after that to it. This is until I get tired.

I feel that with the paper by Heskes and the RNN K-FAC paper that I've read all the pieces for me to start work on RNNs are in place, but I want to give this a while longer. Consider it a vacation from all the programming that I did. The only way I'd ever allow myself to rest from that is if I did something other that would benefit me when I came back to it."