"2:10pm. I've been thinking of how I am going to bring back the MIRNN.
```
                    inm i = matmult (i, input)
                    inm s = matmult (s, state)
                    d2_replicate_activation {
                        fwd=inl (b1,b2,b3,b4) (i,s) -> b1*i*s + b2*s + b3*i + b4 |> fwd
                        bck_in=inl (b1,b2,b3,b4) (i,s) out -> (i*s, s, i, one) |> Tuple.map ((*) (bck out))
                        bck_in'=inl (b1,b2,b3,b4) (i,s) out -> (b1*s+b3, b1*i+b2) |> Tuple.map ((*) (bck out))
                        } (b1,b2,b3,b4) (i,s)
                >>= inl x -> succ (x,x)
```

It is not as easy as it seems. I do not have a good way of whitening the stuff in the activation just yet. I'd have to make different kinds of whitening activations. This is annoying. What I am going to have to do is bring in a different MIRNN. The original one. It will play better with what I have on hand.

This means putting yet more work into optimizing this crap.

2:20pm. Look, for once how about I just not do it and go straight for the prize? I've already skipped learning Bayesian ML today. Why don't I cut my effort short and do differentiable plasticity directly? The online version of it should be really easy. And even though I said I would do Rudder, it is not like I am obligated to do it. Who knows, maybe Zap updates will absolutely improve things well beyond what is expected. I should just make the choices that are most rational with the evidence available.

2:25pm. Now, I need to gather my will. I've decided to go with differentiable plasticity. I already know how good MIRNNs are so there is no need to duplicate that effort. I do not even need to do the char RNN. I can just go straight into poker with them.

Rather than waste 3 weeks doing a broad coverage, I can just dive into whatever I feel like doing. That is Spiral's power.

This is not February and March where I was trying to close where I was in late 2016. I am well beyond that now.

Let me take a break and then I will read the differentiable plasticity paper again."