I finally got it to run.
```
The learning_rate is 2 ** -11.25
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 277 and 723 out of 1000.
The time was 00:00:02.5765934 for: iteration 0
Starting timing for: iteration 1
Winrate is 393 and 607 out of 1000.
The time was 00:00:03.5585593 for: iteration 1
Starting timing for: iteration 2
Winrate is 473 and 527 out of 1000.
The time was 00:00:03.3432389 for: iteration 2
Starting timing for: iteration 3
Winrate is 516 and 484 out of 1000.
The time was 00:00:03.4399500 for: iteration 3
Starting timing for: iteration 4
Winrate is 546 and 454 out of 1000.
The time was 00:00:03.4183568 for: iteration 4
Starting timing for: iteration 5
Winrate is 610 and 390 out of 1000.
The time was 00:00:03.5666894 for: iteration 5
Starting timing for: iteration 6
Winrate is 621 and 379 out of 1000.
The time was 00:00:03.3971819 for: iteration 6
Starting timing for: iteration 7
Winrate is 608 and 392 out of 1000.
The time was 00:00:03.2668097 for: iteration 7
Starting timing for: iteration 8
Winrate is 603 and 397 out of 1000.
The time was 00:00:03.3956248 for: iteration 8
Starting timing for: iteration 9
Winrate is 609 and 391 out of 1000.
The time was 00:00:03.0842113 for: iteration 9
Starting timing for: iteration 10
Winrate is 624 and 376 out of 1000.
The time was 00:00:02.9952256 for: iteration 10
Starting timing for: iteration 11
Winrate is 629 and 371 out of 1000.
The time was 00:00:02.7705989 for: iteration 11
Starting timing for: iteration 12
Winrate is 613 and 387 out of 1000.
The time was 00:00:02.6969283 for: iteration 12
Starting timing for: iteration 13
Winrate is 632 and 368 out of 1000.
The time was 00:00:02.6602538 for: iteration 13
Starting timing for: iteration 14
Winrate is 644 and 356 out of 1000.
The time was 00:00:02.6447747 for: iteration 14
```

It starts off decent, but it definitely has convergence issues.

I want to try the same with an AC.