Here it is.
```
The learning_rate is 2 ** -11.25
------
The CudaRandom pseudorandom seed is 0
Starting timing for: iteration 0
Winrate is 250 and 750 out of 1000.
The time was 00:00:03.1208333 for: iteration 0
Starting timing for: iteration 1
Winrate is 385 and 615 out of 1000.
The time was 00:00:04.4389989 for: iteration 1
Starting timing for: iteration 2
Winrate is 499 and 501 out of 1000.
The time was 00:00:04.5722370 for: iteration 2
Starting timing for: iteration 3
Winrate is 554 and 446 out of 1000.
The time was 00:00:04.5356933 for: iteration 3
Starting timing for: iteration 4
Winrate is 592 and 408 out of 1000.
The time was 00:00:04.7360220 for: iteration 4
Starting timing for: iteration 5
Winrate is 597 and 403 out of 1000.
The time was 00:00:04.8782981 for: iteration 5
Starting timing for: iteration 6
Winrate is 643 and 357 out of 1000.
The time was 00:00:04.4449391 for: iteration 6
Starting timing for: iteration 7
Winrate is 637 and 363 out of 1000.
The time was 00:00:04.8635160 for: iteration 7
Starting timing for: iteration 8
Winrate is 682 and 318 out of 1000.
The time was 00:00:05.1079866 for: iteration 8
Starting timing for: iteration 9
Winrate is 680 and 320 out of 1000.
The time was 00:00:05.2571827 for: iteration 9
Starting timing for: iteration 10
Winrate is 684 and 316 out of 1000.
The time was 00:00:05.0084117 for: iteration 10
Starting timing for: iteration 11
Winrate is 694 and 306 out of 1000.
The time was 00:00:04.9558569 for: iteration 11
Starting timing for: iteration 12
Winrate is 692 and 308 out of 1000.
The time was 00:00:04.6885974 for: iteration 12
Starting timing for: iteration 13
Winrate is 730 and 270 out of 1000.
The time was 00:00:04.4434959 for: iteration 13
Starting timing for: iteration 14
Winrate is 717 and 283 out of 1000.
The time was 00:00:04.4537004 for: iteration 14
```

It is true what Levine said about noise hurting more the better the policy gets. Getting optimal performance is really, really important to me in poker as it will make the difference between a winning and a losing player. Though I think I could improve on the above by using layer norm.

Now that I have this, my goal will be to make a DP layer and plug it into the AC agent. If that succeeds without destabilizing the net then will I have won. I do not actually expect it to improve learning any against the rules based player, but when self play time comes around it will make a marked difference. Against real life players too. The feedforward net simply lacks the capacity to adapt in real time.

Also based on the above, it is really good to see that PRONG is costing me only ~2.3x in computation time. For the benefit it gives that is more than a welcome tradeoff.

https://arxiv.org/abs/1805.09092
Excitation Dropout: Encouraging Plasticity in Deep Neural Networks

Let me read this paper. It is the only paper citing the DP one so far. I was wondering if there is a follow up since it has been a few months since it was last released. Nothing yet.

Since it is getting late, I am not sure I will feel like starting work on DP today, but I'll definitely finish the online variant tomorrow. I've just about gathered enough will to really let it loose.