Hmmm...I keep thinking about the gradient centering part from EigenNets paper. What if I put the critic directly into the whiten function and rather than just do the centering with respect to weights, I propagate it as in a real critic? That might give me better results.
This would have the great benefit of allowing me to experiment with it in the lower layers.

Right now propagating rewards is such a pain in the ass as I need special cases when it is on the host and on the device.

Well, forget that for now. There will be time for tightening up the design later. Right now I need to focus on differentiable plasticity.

Let me sketch it out.