"11:10am.
```
inm x = matmult ({T=in},out)
```

This is how I transpose I think. I think I added this at some point.

11:20am. Aggghhh, this is such a massive pain in the ass. I mean the `n` parameter. I have no good way of doing the global reductions if it were a scalar. It would require various special machinery that I do not have right now, or alternatively hacks.

I am much better off making it its own matrix. This is not how it is in the paper, but maybe it won't be too bad? It can only make the net more flexible.

11:30am.

```
    // #Recurrent
    inl hebb n {in out H} =
        inm x = matmult ({T=in},out)
        map {
            fwd=inl {n x H} -> n * x + (1 - n) * H
            bck={
                n=inl {n x H} _ -> x - H
                x=inl {n x H} _ -> n
                H=inl {n x H} _ -> -n
                }
            } {in out H}
```

First time I used this form for some reason.

```
            inm W =
                match d with
                | {} as state ->
                    inm H = hebb weights.n state
                    hadmultb (weights.input.alpha, H) weights.input.bias
                | _ -> succ weights.input
```

It has been a while since I did this sort of programming. Minus all the type errors, this should be a complete implementation of differentiable plasticity.

It was really trivial, but I am really dragging my feet here. Let me have that break and then I will test this thing. If it runs without destabilizing the AC like the MI net tended to do, then I will have succeeded.

It does not seem much, but a stable RNN architecture for RL is actually one of the most important steps to accomplish. The whole plan will be in jeopardy unless I manage to do this. Differentiable plasticity should work for this purpose, but I need to make sure."