"2:15pm.
```
        apply = inl {d with weights input} s ->
            assert (input.span_outer = 1) "The differentirable plasticity layer supports only online learning for now."
            inl apply =
                inm W, H =
                    match d with
                    | {state} ->
                        inm H = hebb weights.input.n state
                        succ (hadmultb (weights.input.alpha, H) weights.input.bias, H)
                    | _ -> succ (weights.input, ())
                inm out = matmultb (input, W) weights.bias >>= activation
                succ {out state={out input H}}
            inl {out={out state} bck} = apply s
            {out state bck}
```

Let me just do it like this. Also this kind of thing is going to make it really annoying to pass in the learning rate into whiten when the time comes to do it.

Let me go back a little to the prong layer to show what I mean.

```
        apply = inl {weights input} s ->
            inl {out=a bck=bck_a} = whiten config {steps_until_inverse_update learning_rate_modifier} weights input s
            inl {out=b bck=bck_b} = activation a s
            {
            out=b
            bck=inl x -> bck_b(); bck_a x
            }
```

This thing. I am bypassing the usual mechanism just to get it to work.

I just hadn't thought that I would never need `bck` to pass its future inputs to the previous iteration.

2:25pm. Resolving this will require a core change in how the monad works. Rather than composing functions I am going to go to a list based representation. This will allow me to map fold over it. I need this in order to handle state for the feedforward part and I am going to need it for the backwards part. The way I am doing right now is just so ridiculous.

Actually, it is amazing that it took me this long to figure out that this is necessary. That the way I am doing things now is broken should have occurred to me back when I was trying out Q learning for the first time. Back then I had a lot of difficulty figuring out how to pass future values to the present so I opted to do it on the forward pass which was a pretty bad idea, but not a wrong one unfortunately.

This is essentially another bad habit I am dragging into 2018 from 2016 when I did the F# ML library. I used functions for the backward pass there. Map folding them was not something that was even on the horizon back then and it remained fixed it my mind that closures were just the right way to do this.

Truly, I am such a fool, but I can do better this time. I can lift this restriction on myself.

2:45pm. Right. Now let me test out the differentiable plasticity layer on the poker game. I'll do the redesign after that."