" ```             | {constant={dim init}} ->                 inl rec loop = function                     | _ :: x' -> inl x -> loop x'                     | () -> init                 s.CudaKernel.init {dim} (loop type dim) ```
The way `init` is made is really massive pain in the ass. It is one function I should really restructure.

I should split it into `iter`, `iter2` and `iter3` and the corresponding `init`s. There is literally no point to it being like this right now. I did it for those `rev_thread` things, but I literally used that only once. The rest of the time I have to deal with these pain in the ass partially applied parameters which serve no point in being partially applied since everything is done in a single loop anyway.

Along with the redesign on how the backward pass is handled that will also be on my agenda."