Except it is not that easy. The prong layer is in a horrible state. I need to massively clean it up. I need to give it its own initializer. I also need to do that redesign I talked about.
Also I am going to have to do the char-RNN test. I do not trust that differentiable plasticity is doing what it intends to do. It might be only using the feedforward part and I might not even know it.

First let me make a specialized initializer for PRONG. I want to clean it up for the time being.