Though I am starting to get into the stride of this, I think I will stop here for the day. I do not feel like it anymore. I'll leave everything as it is right now, in a completely broken state. Tomorrow I will bring it back all online and do the char RNN. I'll remove the TD learning stuff and put it in a separate text file.
I'll also want to compare diff plasticity to a standard MI-RNN just to know where I stand.

I should unify the actor and the critic into a single layer so I do not collect the statistics and whiten using the inputs to it twice. I need to figure out how I want to design this.

Also I must not forget that I removed the `learning_rate_modifier`.

Actually the easiest way to do it would be to just allow multiple input weight matrices. That would easily get me what I want without much effort. It should be ideal and not much fuss. I am going to have to embedd them into the back layer then. But doing it like that would also give me a succinct answer how to whiten those bias vectors.

...

Ah, I see. Though PRONG as a method is tied to matrix multiplication, the actual function for it does not need to be like that. I can separate it into the forward projection, backward projection and fill the gaps using functions. My idea of how it should be done is just too narrow.

Yeah, that is it. I've been way too impatient lately. Sometimes in order to get thing in proper order and draw out the true power of the program it is simply necessary to spend some time pondering the design of things and to go at it step by step.

Now that I am not under the pressure to finish the testing of anything, I will do that very necessary upkeep. Tomorrow I will generalize the `whiten` function. And maybe start by renaming it to `naturalize`. Natural gradient is really not doing whitening, but the square of it.

Things are going very well. I have roughly half of a good agent now. I just need to go further and seal the deal. Even if that took me a whole month it would be a worthwhile price to pay.