I have so much to do and yet I am slacking. The Kalman Batch Norm paper is honestly interesting, but since the transition parameters it uses cannot be optimized at compile time I am not sure whether that is something that could be used for K-FAC.
Forget that. Maybe I'll find a route to understanding the K-FAC papers eventually, but it not a priority.

Last night I did a lot of thinking about the iter kernels that I will make so I intend to work on that today.

For today, the priority will be to tests back online. This should be easy. After that I will make the new iter kernels. I am not sure how long that will take. Making them will be easy enough, but I am not sure how long the redesign itself will take.

After that I will need to generalize the `naturalize` function.

Let me go for it.