``` met iter w {d with dim} f =     inl l = length dim     inl blockDim = min l <| match d with {num_threads} -> num_threads | _ -> 256     inl gridDim = divup l blockDim     w.run {blockDim gridDim         kernel = cuda grid_for {blockDim gridDim} .x dim {body=inl {i} -> f i}         } ```
Without a doubt this is what an `iter` function should be. Just compare it to the one before.

```
/// Iterates over the dimensions.
/// Takes in the optional {thread_limit} or {rev_thread_limit} as the first argument in order to control the degree of parallelism.
met iter w {d with dim} f =
    inl dim = HostTensor.map_dims dim
    inl rec merge = function
        | thread_limit :: l', dim :: d' -> {dim thread_limit=min thread_limit (s dim)} :: merge (l', d')
        | (), d' -> Tuple.map (inl dim -> {dim thread_limit=()}) d'
    inl d =
        match d with
        | {thread_limit} -> merge (Tuple.wrap thread_limit,dim)
        | {rev_thread_limit} -> merge (Tuple.wrap rev_thread_limit, Tuple.rev dim) |> Tuple.rev
        | _ -> merge ((),dim)
    inl s = function {thread_limit=() dim} -> s dim | {thread_limit} -> thread_limit
    inl near_to = Tuple.foldl (inl a (!s b) -> a*b) 1 d
    inl blockDim = min near_to 256
    inl gridDim = divup near_to blockDim

    w.run {blockDim gridDim
        kernel = cuda
            grid_for {blockDim gridDim} .x {from=0; near_to} {body=inl {i} ->
                inl l,_ = Tuple.foldr (inl ((!s x_span) & x) (l,i) -> (i % x_span + x.dim.from) :: l, i / x_span) d ((),i)
                inl rec loop f = function
                    | {thread_limit=()} :: d', i :: i' -> loop (f i) (d', i')
                    | {thread_limit=by dim={near_to}} :: d', from :: i' -> forcd {from by near_to body=inl {i} -> loop (f i) (d',i')}
                    | (), () -> f; () // The output of f should be unit.
                loop f (d,l)
                }
        }
```

It has so much shit I will never ever use. The only I used the `rev_thread_limit` thing is during the preprocesing of the tiny shakespeare dataset and there it was not even necessary.

Allowing the regular loop to iterate over arbitrary dimensions was trully a master stroke.

Also the thing that really annoys me about iter is the partial application of index argument. This is good when they are actually separate loops, but in the function above they definitely are not unless you pass it those `thread_limit` arguments.

Apart from iters, I will redo the `d2_redo` so the index arguments it gives out are not inverted. That is a really, really bad design choice. I will do a bunch of renaming so what the function does is better reflected in it.