Let me make `iter` the very first kernel as it should be rather than having it be at the end of the module.
```
met map' w f in out =
    inl in, out = zip in, zip out
    assert (in.dim = out.dim) "The input and output dimensions must be equal."
    inl in = flatten in |> to_dev_tensor
    inl out = flatten out |> to_dev_tensor
    inl in_a :: () = in.dim

    inl blockDim = 128
    inl gridDim = min 64 (divup (s in_a) blockDim)

    w.run {
        blockDim gridDim
        kernel = cuda // Lexical scoping rocks.
            grid_for {blockDim gridDim} .x in_a {body=inl {i} ->
                inl out = out i
                inl in = in i
                out .set (f in.get out.get)
                }
        }
```

Ah, `map'`. To make this kernel the first time again...

It is difficult to explain just how much difference having access to it in a ML library speeds things up. Those who haven't suffered for a year making a ML library in a lesser language won't be able to understand it.

https://www.reddit.com/r/DotA2/comments/99idug/the_international_8_openai/?st=jl6hnzha&sh=066b3e0b
https://news.ycombinator.com/item?id=17823286

I'll read these comments later. Right now I need to focus.

I've moved the `// Lexical scoping rocks.` comment to `iter`. Fundamentally, that comment needs to be in the module somewhere and no better place than the very first kernel.

Now...since I really do not need map to do flattening, I'll remove all this stuff and reimplement it in terms of `init`. I'll also add `mapi`. I wanted that a few times, but could not get it which was pretty annoying.