```         /// Applies the tensor. `i` can be a tuple.         apply = inl data i ->             Tuple.foldl (inl {data with dim} i ->                 match dim with                 | () -> error_type "Cannot apply the tensor anymore."                 | {from near_to} :: dim ->                     assert (i >= from && i < near_to) "Argument out of bounds."                     {data with bodies = Struct.map (inl ar -> tensor_apply ar (i-from)) self; dim}                     |> facade                 ) data (Tuple.wrap i) ```
Ok, that is one more test down. This will simplify things a bit. 14 tests to go.

I think that I am going to pare down the `map_redo` functions. It is honestly a major pain in the ass to have to remember to pass in two arguments. For adding biases, I'll just use `init_redo` and let the magic of lexical scoping taking care of the rest.

The whole module is overengineered and underengineered in places at the same time.