"6:05pm. Done.
6:20pm. Ah...I have a really, really good idea on how to make infering types for these kernels much easier.

I am going to introduce expections on the type level. Rather than making custom inference functions for all that stuff, I will just run the kernel normally and throw an exception in outit. That is going to cut the amount of code I need to write to get this to work by a serious amount. It is going to cause memory to waste and compile times to go up, but who cares about that. It is going to cut the amount of programming work by half and that is what matters.

6:35pm. Ohhh...this might be it. Time might really be unfreezing for me again. I thought I would not find a way to ever make Spiral fundamentally better, but this idea would automate just so much in the CudaKernel module.

Maybe at some point I will even figure out how to make those exceptions resumable which would give me a more generalized way of doing type inference in a language with first class types like Spiral.

6:40pm. No, actually forget it.

Actually, the main reason why I have so much work is because the module is so convoluted.

What I will do tomorrow is take out those map functions and restrict the whole modules so every function has a strict `{init outit}` interface. That will make everything else much easier. If I think about it that would reduce inference to just passing in dimnesions and then piping them through those two functions. Everything else is some kind of scan or reduce anyway.

Really, all parallel programming boils down to map and fold of some sort. If it does not then it is probably some kind of mistake in the algorithm. Maybe filter too, but I've yet to run into that one.

Right, so here is the plan for tomorrow.

1) I will restructure the CudaKernel module so every single kernel only takes in an init and an outit for inputs and outputs.
2) I will use that structure to automate type inference for when only input is given.
3) Make the interface functions.

Honestly, most of that work that I did for 3 was a waste of time. I would have been much better taking a principled design stance from the start, but oh well.

Things like that happen. Earlier in the year Spiral still was not in top shape and I had only really begun work on the kernels at that point. I knew myself that I did not know exactly what I was doing.

Well, admittedly some functions might not fit into the exact framework. Maybe I'll go with the exception throwing idea in the end. Forget that for now.

I am done for the day here. Let me watch the Danganronpa 3 playthrough."