"1:20pm.
```
Edit: If `h` is 1x1 then it definitely does not make sense as that would only rescale the matrix. Rather than resulting in a non-symmetric steady state matrix, it would result in a symmetric covariance matrix so it cannot be true.

Maybe if a batch learning context is assumed then it would be possible to get somewhere with this as the `h` matrix would be full rank?
```

1:45pm. No, I just have no idea. Forget it. I really need better math skills to crack this. I should focus on chiseling down the Cuda kernels instead.

Let me resume what I am supposed to be doing."