"6:40pm. It seems I spent over 3h in bed just thinking. I am done with lunch and will call it a day here. But before that let me write some things down.
```
/// Maps the input and then reduces twice, first along the inner dimension and then along the middle.
met init_redo_redo w {d with dim=c,b,a init outit redo_in redo_mid} =
    assert (dim_in_a = out_a) "Input's outermost and output's dimension must be equal."
    inl x = lit_min 1024 (length a)
    inl y = lit_min (1024 / x) (length b)
    w.run {
        blockDim={y x}
        gridDim={z=min 64 (length c); y=min 64 (divup (length b) y)}
        kernel = cuda
            inl grid_for = grid_for {blockDim gridDim}
            grid_for .z c {body=inl {i} ->
                inl {mapi_in redo neutral_elem} = redo_mid
                inl x =
                    inl in = in i
                    grid_for .y b {state=dyn neutral_elem; body=inl {state i=j} ->
                        inl x =
                            inl in = in j
                            inl {d with redo neutral_elem mapi_in} = redo_in
                            grid_for .x a {state=dyn neutral_elem; body=inl {state i=k} ->
                                inl in = in k .get
                                redo state (mapi_in i j k in)
                                }
                            |> block_reduce_2d (blockDim.y,threadIdx.y) (blockDim.x, threadIdx.x) redo
                        redo state (mapi_in i j x)
                        }

                if threadIdx.x = 0 then
                    inl x = block_reduce_1d (blockDim.y, threadIdx.y) redo x
                    if threadIdx.y = 0 then
                        inl out = out i
                        out.set (map_out x out.get)
                }
        }
```

This function tricked me. I spent a lot of time thinking of how I should increse the flexibility of the CudaKernel module functions, but only very recently it occured to me that I do not actually need this function.

The reason I made it originally was to sum up all the individual Q values and then take the max over them. This was back when I was testing distributional Q learning.

Now rather than reduce two separate dimensions, what would have been better would be to make a kernel with a `redo` like this.

```
redo = inl a b -> if a.i = b.i then {a with x=a.x+b.x} elif a.x > b.x then a else b
```

Pretty much all sequential multiple reductions can be decomposed into this pattern. The reason why I did not use the standard reduction kernel is because it was intended for 2d tensors only. This is unlike the one I have now which supports arbitrary dimensions.

I actually did not even have this in mind when I was doing it, but now I can clearly see that I released quite a bit of power when I did this. Without meaning to, I completely obsoleted the `dredo` kernel. I never even thought about doing it like this last time, but now it came so naturally once the ability to do so was there. Maybe there is a lesson there - the tendency is to not think about the impossible.

I also though about what the general philosophy of the library is going to be. I think that for map,scan,redo,init_redo,init_scan,redo_init,scan_init and other simple 1d and 2d kernels it would be fine if I continue with my present course. I'll give them their own allocators in order to make them ergonomic to use.

But for functions like...`init_d1_seq_broadcast` and `iteri_dd1_seq_broadcast` and `mapi_d1_inscan_mapi_d1_reduce_mapi`.

The kind of functions that are to be used to make optimized implentations, I am going to get rid of the later two and turn the first one into a DSL. I'll also get rid of the allocator version of it since I never actually used that variant. The way layer norm is implemented is attrocious and I could have done much, much better.

It should have been done in a monadic fashion.

Then rather than getting the variables from a tuple, I'll do all the operations where they should have been. Those optimizer functions really need flexibility and I made the mistake of trying to make them ergonomic instead.

Let me stop here for the day. I'll continue this tomorrow."